+++ elfo-core/src/dumping/mod.rs
@@ -6,7 +6,7 @@ use serde::{Serialize, Serializer};
 #[cfg(feature = "unstable")] // TODO: patch `stability`, again.
 pub use self::{
     control::{CheckResult, DumpingControl},
-    dump::{Direction, Dump, ErasedMessage, MessageKind, MessageName, Timestamp},
+    dump::{Direction, Dump, ErasedMessage, MessageKind, MessageName},
     dumper::Dumper,
     extract_name::{extract_name, extract_name_by_type},
     raw::Raw,